### 2021-01-26 #2981 @jedbrown
I can take this if Jack is conflicted (though I can't promise to be fast -- I'm pretty swamped).

@andreasmang I don't see any automated tests and the docs seem hard to navigate (though I'm not the target audience) to understand how to prepare data, assess quality, and contribute. Please consider adding GitHub Actions or other automated tests, as well as [recommended community docs](https://github.com/andreasmang/claire/community).
