### 2021-01-31 4b286af
working ncc gpu implementation
