### 2021-02-01 #2981 @andreasmang
> @andreasmang please look through your paper as it is shown above and correct the references. Many have parentheses when they should not.

@kthying Sorry for the late response. I was assuming I get notified by email once there is any action required. I have fixed the references in the CLAIRE repo.
