### 2021-02-01 #2981 @andreasmang
> I can take this if Jack is conflicted (though I can't promise to be fast -- I'm pretty swamped).
>
> @andreasmang I don't see any automated tests and the docs seem hard to navigate (though I'm not the target audience) to understand how to prepare data, assess quality, and contribute. Please consider adding GitHub Actions or other automated tests, as well as [recommended community docs](https://github.com/andreasmang/claire/community).

@jedbrown Thanks for your recommendation. I will look into your suggestions and update the documentation. I'll ping you once I believe we have addressed everything you have brought to my attention...
