### 2021-02-02 56a515c
mean subtracted NCC implementation - not converging
