+++ src/DistanceMeasure/DistanceMeasureNCC.cpp
@@ -69,9 +69,6 @@ PetscErrorCode DistanceMeasureNCC::ClearMemory() {
     PetscFunctionReturn(ierr);
 }
 
-
-
-
 /********************************************************************
 * @brief Set up scale for NCC measure given by
 * scale = D_SL2(mR,mT)/D_NCC(mR,mT)
@@ -80,8 +77,8 @@ PetscErrorCode DistanceMeasureNCC::SetupScale(){
     PetscErrorCode ierr = 0;	
     DistanceMeasureKernel::EvaluateFunctionalNCC kernel;
     ScalarType *p_mr = NULL, *p_mt = NULL, *p_w = NULL;
-    IntType nc;
-    ScalarType norm_l2, norm_mT, norm_mR, inpr_mT_mR, mTi, mRi;
+    IntType nc, ng;
+    ScalarType norm_l2, norm_mT, norm_mR, inpr_mT_mR, sum_mT, sum_mR;
     int rval;
     ScalarType l2distance, nccdistance, hd;
 
@@ -97,6 +94,7 @@ PetscErrorCode DistanceMeasureNCC::SetupScale(){
     kernel.nc = this->m_Opt->m_Domain.nc;
     nc = this->m_Opt->m_Domain.nc;
     kernel.nl = this->m_Opt->m_Domain.nl;
+    ng = this->m_Opt->m_Domain.ng;
     hd  = this->m_Opt->GetLebesgueMeasure();   
 
     ierr = this->m_TemplateImage->GetArrayRead(kernel.pMt); CHKERRQ(ierr);
@@ -125,16 +123,28 @@ PetscErrorCode DistanceMeasureNCC::SetupScale(){
 
     rval = MPI_Allreduce(&kernel.inpr_mT_mR_loc, &inpr_mT_mR, 1, MPIU_REAL, MPI_SUM, PETSC_COMM_WORLD);
     ierr = Assert(rval == MPI_SUCCESS, "mpi error"); CHKERRQ(ierr);
+
+    rval = MPI_Allreduce(&kernel.sum_mR_loc, &sum_mR, 1, MPIU_REAL, MPI_SUM, PETSC_COMM_WORLD);
+    ierr = Assert(rval == MPI_SUCCESS, "mpi error"); CHKERRQ(ierr);
+    
+    rval = MPI_Allreduce(&kernel.sum_mT_loc, &sum_mT, 1, MPIU_REAL, MPI_SUM, PETSC_COMM_WORLD);
+    ierr = Assert(rval == MPI_SUCCESS, "mpi error"); CHKERRQ(ierr);
+
     
     ierr = this->m_ReferenceImage->RestoreArray(); CHKERRQ(ierr);
     ierr = this->m_TemplateImage->RestoreArray(); CHKERRQ(ierr);
     ierr = RestoreRawPointerRead(this->m_ObjWts, &kernel.pWts); CHKERRQ(ierr);
 
+    // update the inner products and norms to include the image mean differences
+    inpr_mT_mR = inpr_mT_mR - sum_mR*sum_mT/ng;
+    norm_mT = norm_mT - sum_mT*sum_mT/ng;
+    norm_mR = norm_mR - sum_mR*sum_mR/ng;
+
     // Calculate two distance measures and scale
     l2distance = 0.5*hd*norm_l2/static_cast<ScalarType>(nc);
     nccdistance = 0.5 - 0.5*(inpr_mT_mR*inpr_mT_mR)/(norm_mT*norm_mR);
     this->m_Opt->m_Distance.scale = l2distance/nccdistance;
[…]
