### 2021-02-02 #2981 @kthyng
> @kthying Sorry for the late response. I was assuming I get notified by email once there is any action required. I have fixed the references in the CLAIRE repo.

You'll just be getting github pings, so I'd recommend changing your github email notifications so that you are emailed when you are pinged here.
