### 2021-02-02 #2981 @andreasmang
> > @kthying Sorry for the late response. I was assuming I get notified by email once there is any action required. I have fixed the references in the CLAIRE repo.
>
> You'll just be getting github pings, so I'd recommend changing your github email notifications so that you are emailed when you are pinged here.

@kthyng Thanks for pointing this out. The problem was that notifications went to the address of my old affiliation (which is no longer active). No I should be good to go...
