### 2021-02-04 #2981 @jedbrown
I'd love to, but I'm just ridiculously overcommitted. Today was like a parody of two-parent work-from-home with sick kids and my experience of the past 11 months is that it won't get better in the near future so I'm afraid I have to decline at this time. If nobody else is available, I can reassess in a month or two.
