### 2021-02-05 505d2ed
Create CODE_OF_CONDUCT.md
