### 2021-02-05 fb8dd3a
Create CODE_OF_CONDUCT.md
