+++ CODE_OF_CONDUCT.md
@@ -55,8 +55,8 @@ further defined and clarified by project maintainers.
 ## Enforcement
 
 Instances of abusive, harassing, or otherwise unacceptable behavior may be
-reported by contacting the project team at [Andreas Mang](mailto:andreas@math.uh.edu).
-All complaints will be reviewed and investigated and will result in a response that
+reported by contacting the project team at andreas@math.uh.edu. All
+complaints will be reviewed and investigated and will result in a response that
 is deemed necessary and appropriate to the circumstances. The project team is
 obligated to maintain confidentiality with regard to the reporter of an incident.
 Further details of specific enforcement policies may be posted separately.
