+++ doc/README-NEWS.md
@@ -1,4 +1,4 @@
 # CLAIRE: News Feed
 
-## 2018
+*28/2020 We have released a GPU version of CLAIRE. If you are interested in using our new (multi-node multi-)GPU version, switch to the **GPU branch**. If you are interested in learning more about the GPU version of CLAIRE, check out our [publications](doc/README-REFERENCES.md).
 * 11/2018 Support for Normalized-Cross-Correlation (NCC) has been added to CLAIRE.
