+++ doc/README-REFERENCES.md
@@ -7,13 +7,13 @@ A. Mang, A. Gholami, C. Davatzikos & G. Biros. *CLAIRE: A distributed-memory sol
 A BibTeX entry for LaTeX users is
 ```TeX
 @article{Mang:2019a,
-author = {A. Mang and A. Gholami and C. Davatzikos and G. Biros},
-journal = {SIAM Journal on Scientific Computing},
-number = {5},
-pages = {C548--C584},
-title = {{CLAIRE}: A distributed-memory solver for constrained large deformation diffeomorphic image registration},
-volume = {41},
-year = {2019}}
+author = "A. Mang and A. Gholami and C. Davatzikos and G. Biros",
+journal = "SIAM Journal on Scientific Computing",
+number = "5",
+pages = "C548--C584",
+title = "{CLAIRE}: A distributed-memory solver for constrained large deformation diffeomorphic image registration",
+volume = "41",
+year = "2019"}
 ```
 
 
