### 2021-02-05 2e45237
readme (installation guide) updated
