+++ doc/README-INSTALL.md
@@ -1,20 +1,48 @@
-# CLAIRE: Detailed Installation Guide
+# Installation and Requirements
 
+Go back to (README.md)[README.md].
 
 ## Content
 
-* [Requirements](#requirements)
-* [Dependencies](#dependencies)
-	* Required Dependencies
-  * Step 1: Downloading Dependencies
-  * Step 2: Installing Dependencies
-  	* Quick Shot
-  	* Detailed Instructions
-  * Step 3: Setting Environment Variables
-* [Building CLAIRE](#buildclaire)
-* [Additional Info for Dependencies](#depsinf)
-* [Troubleshooting](#faq)
+* [Installation Overview](#installation)
+* [Detailed Installation Guide](#verboseinstall)
+	* [Requirements](#requirements)
+	* [Dependencies](#dependencies)
+	* [Building CLAIRE](#buildclaire)
 
+## Installation Overview <a name="installation"></a>
+
+What follows is a minimal installation guide. We provide scripts to download and install the dependencies using generic settings that have worked on our systems. A more detailed installation guide can be found [below](#verboseinstall).
+
+To install the dependencies go to the top level directory of CLAIRE in your command window and execute the following commands in your command window:
+```bash
+cd deps
+./get_libs.sh
+./build_libs.sh --build
+```
+
+The first bash script downloads the libraries you need to install. The second bash script compiles these libraries using default settings. To add these dependencies to your environment type the following into your command line and press return:
+```bash
+source environment_vars.sh
+```
+
+This step needs to be done every time you log out of your computer. As an alternative, you can add the content of `environment_vars.sh` to your `.bashrc` or `bash_profile`.
+
+Return to the top level directory of CLAIRE and type
+```bash
+cd ..
+make -j
+```
+
+If you would like to verify if CLAIRE has been installed correctly run the following command in your command window:
+
+```bash
+./bin/claire -synthetic 0
+```
+
+If you have trouble executing CLAIRE take a look at the [examples and tests](README-RUNME.md) described in [README-RUNME.md](README-RUNME.md).
+
+# Detailed Installation Guide <a name="verboseinstall"></a>
 
 ## Requirements <a name="requirements"></a>
 
@@ -22,20 +50,37 @@
 * cmake ([https://cmake.org](https://cmake.org)); required by `AccFFT` and `niftilib`
 * python ([https://www.python.org](https://www.python.org)); version 2.7; required by `PETSc`
 
-
 Make sure that the standard *wrappers* for `mpicc` and `mpicxx` are available on your system (either by loading the appropriate modules and/or by setting up the appropriate `PATH` and `LD_LIBRARY_PATH` definitions below). The compilation has been tested with `Open MPI`, `MVAPICH`, and `Intel MPI`.
-
 
 ## Dependencies <a name="dependencies"></a>
 
-### Required Dependencies
 CLAIRE requires the following libraries to be installed on your system:
 * `FFTW` [http://www.fftw.org](http://www.fftw.org)
+	* file: [fftw-3.3.6-pl2.tar.gz](http://www.fftw.org/fftw-3.3.6-pl2.tar.gz)
+	* description: library for computing FFTs
+	* additional details can be found at [http://www.fftw.org](http://www.fftw.org)
 * `AccFFT` [http://accfft.org](http://accfft.org)
[…]
