### 2021-02-05 645b7b7
readme (installation guide) updated
