### 2021-02-05 c2767dd
readme (installation guide) updated
