### 2021-02-05 c1cc366
readme (installation guide) updated
