+++ doc/README-INSTALL.md
@@ -1,6 +1,6 @@
 # Installation and Requirements
 
-Go back to [README.md](README.md).
+Go back to [README.md](../README.md).
 
 ## Content
 
