+++ doc/README-REFERENCES.md
@@ -1,5 +1,7 @@
 # References
 
+Go back to [README.md](../README.md).
+
 If you plan on using CLAIRE in your research please cite the following manuscript:
 
 A. Mang, A. Gholami, C. Davatzikos & G. Biros. *CLAIRE: A distributed-memory solver for constrained large deformation diffeomorphic image registration*. SIAM Journal on Scientific Computing 41(5):C548--C584, 2019 [[arxiv](https://arxiv.org/abs/1808.04487), [sisc](https://epubs.siam.org/doi/abs/10.1137/18M1207818)].
