### 2021-02-05 e23af8b
readme (installation guide) updated
