+++ doc/README-INSTALL.md
@@ -3,12 +3,12 @@
 Go back to [README.md](../README.md).
 
 ## Content
-
 * [Installation Overview](#installation)
 * [Detailed Installation Guide](#verboseinstall)
 	* [Requirements](#requirements)
 	* [Dependencies](#dependencies)
 	* [Building CLAIRE](#buildclaire)
+* [Instructions for Specific Systems](#systems)
 * [Trouble Shooting](#faq)
 
 ## Installation Overview <a name="installation"></a>
@@ -146,6 +146,34 @@ make -j
 ```
 
 If you build in parallel using `make -j`, on certain systems to many threads will be used. This will result in compilation errors. To fix this, run `make -j 12` instead (for quick access, you may want to define an alias in your `~/.bashrc`).
+
+
+## Instructions for Specific Systems <a name="systems"></a>
+
+Information about the Texas Advanced Computing Center ([TACC](https://www.tacc.utexas.edu)) can be found at [https://www.tacc.utexas.edu](https://www.tacc.utexas.edu).In the following we will mainly provide information on which modules you need to load on the individual TACC systems to compile the code and the libraries. For more information on the systems please visit the [TACC](https://www.tacc.utexas.edu) web pages.
+
+
+## Maverick Systems
+
+Info about Maverick can be found here: [https://www.tacc.utexas.edu/systems/maverick](https://www.tacc.utexas.edu/systems/maverick).
+
+CLAIRE has been tested with the following module settings:
+
+```bash
+module load intel/15.0.3
+module load impi/5.0.3
+module load git/2.7.0
+module load cmake/2.8.12.2
+```
+
+
+
+## Stampede
+
+```bash
+module load intel/14.0.1.106
+module load impi/5.0.3 
+```
 
 
 ## Troubleshooting <a name="faq"></a>
