+++ doc/README-TACC.md
@@ -1,28 +1,0 @@
-# TACC Systems
-
-Go back to [README.md](../README.md).
-
-Information about the Texas Advanced Computing Center ([TACC](https://www.tacc.utexas.edu)) can be found at [https://www.tacc.utexas.edu](https://www.tacc.utexas.edu).In the following we will mainly provide information on which modules you need to load on the individual TACC systems to compile the code and the libraries. For more information on the systems please visit the [TACC](https://www.tacc.utexas.edu) web pages.
-
-
-## Maverick Systems
-
-Info about Maverick can be found here: [https://www.tacc.utexas.edu/systems/maverick](https://www.tacc.utexas.edu/systems/maverick).
-
-CLAIRE has been tested with the following module settings:
-
-```bash
-module load intel/15.0.3
-module load impi/5.0.3
-module load git/2.7.0
-module load cmake/2.8.12.2
-```
-
-
-
-## Stampede
-
-```bash
-module load intel/14.0.1.106
-module load impi/5.0.3 
-```
