### 2021-02-05 7c9d486
readme (installation guide) updated
