+++ doc/README-INSTALL.md
@@ -153,7 +153,7 @@ If you build in parallel using `make -j`, on certain systems to many threads wil
 Information about the Texas Advanced Computing Center ([TACC](https://www.tacc.utexas.edu)) can be found at [https://www.tacc.utexas.edu](https://www.tacc.utexas.edu).In the following we will mainly provide information on which modules you need to load on the individual TACC systems to compile the code and the libraries. For more information on the systems please visit the [TACC](https://www.tacc.utexas.edu) web pages.
 
 
-## Maverick Systems
+### Maverick Systems (TACC)
 
 Info about Maverick can be found here: [https://www.tacc.utexas.edu/systems/maverick](https://www.tacc.utexas.edu/systems/maverick).
 
@@ -168,7 +168,7 @@ module load cmake/2.8.12.2
 
 
 
-## Stampede
+### Stampede (TACC)
 
 ```bash
 module load intel/14.0.1.106
