+++ src/Solver/TransportEquationSL.cpp
@@ -57,6 +57,8 @@ PetscErrorCode TransportEquationSL::Initialize() {
     
     this->m_SemiLagrangianMethod = nullptr;
     this->m_GradientState = nullptr;
+    this->update_grad = false;
+    this->update_gradx = false;
 
     PetscFunctionReturn(0);
 }
@@ -114,14 +116,17 @@ PetscErrorCode TransportEquationSL::SolveForwardProblem() {
         ierr = this->SolveStateEquation(this->m_VelocityField); CHKERRQ(ierr);
     }
     
-    ierr = this->ComputeGradientState(VelocityIsZero); CHKERRQ(ierr);
+    update_grad = true;
+    update_gradx = true;
+    
+    //ierr = this->ComputeGradientState(VelocityIsZero, true, true); CHKERRQ(ierr);
     
     this->m_Opt->Exit(__func__);
 
     PetscFunctionReturn(ierr);
 }
 
-PetscErrorCode TransportEquationSL::ComputeGradientState (bool V0) {
+PetscErrorCode TransportEquationSL::ComputeGradientState (bool V0, bool grad, bool gradx) {
   PetscErrorCode ierr = 0;
   IntType nt, nc;
   const ScalarType *pM = nullptr;
@@ -136,68 +141,74 @@ PetscErrorCode TransportEquationSL::ComputeGradientState (bool V0) {
   ierr = Assert(this->m_VelocityField != nullptr, "null pointer"); CHKERRQ(ierr);
   ierr = Assert(this->m_WorkVecField[0] != nullptr, "null pointer"); CHKERRQ(ierr);
     
-  if (V0) {
-    if (this->m_GradientState) {
-      for (IntType k = 0; k < nc; ++k) {
-        ierr = this->m_StateVariable->GetArrayRead(pM, k); CHKERRQ(ierr);
-        ierr = this->m_Differentiation->Gradient(this->m_GradientState[k], pM); CHKERRQ(ierr);
-        if (this->m_GradientXState) {
-          ierr = this->m_GradientXState[k]->Copy(this->m_GradientState[k]); CHKERRQ(ierr);
-        }
-      }
-      ierr = this->m_StateVariable->RestoreArray(); CHKERRQ(ierr);
-      for (IntType j = 1; j < (this->m_Opt->m_RegFlags.runinversion?nt+1:1); ++j) {
+    if (V0) {
+      if (this->m_GradientState && grad && update_grad) {
         for (IntType k = 0; k < nc; ++k) {
-          this->m_GradientState[j*nc+k]->Copy(this->m_GradientState[k]);
-          if (this->m_GradientXState && j < nt) {
-            ierr = this->m_GradientXState[j*nc + k]->Copy(this->m_GradientState[k]); CHKERRQ(ierr);
+          ierr = this->m_StateVariable->GetArrayRead(pM, k); CHKERRQ(ierr);
+          ierr = this->m_Differentiation->Gradient(this->m_GradientState[k], pM); CHKERRQ(ierr);
+          if (this->m_GradientXState) {
+            ierr = this->m_GradientXState[k]->Copy(this->m_GradientState[k]); CHKERRQ(ierr);
           }
         }
-      }
-    } else if (this->m_GradientXState) {
-      for (IntType k = 0; k < nc; ++k) {
-        ierr = this->m_StateVariable->GetArrayRead(pM, k); CHKERRQ(ierr);
-        ierr = this->m_Differentiation->Gradient(this->m_WorkVecField[0], pM);
-        for (IntType j = 0; j < (this->m_Opt->m_RegFlags.runinversion?nt:1); ++j) {
[…]
