### 2021-02-13 #2981 @mjsottile
:wave: @NMontanaBrown Would you be interested in providing a review for this JOSS submission?  It appears to be in the vicinity of your areas of expertise.
