### 2021-02-13 #3038 @whedon
[REVIEW]: CLAIRE: Constrained Large Deformation Diffeomorphic Image Registration on Parallel Computing Architectures
**Submitting author:** @andreasmang (<a href="http://orcid.org/0000-0003-4605-3290">Andreas Mang</a>)
**Repository:** <a href="https://github.com/andreasmang/claire" target ="_blank">https://github.com/andreasmang/claire</a>
**Version:** v0.08
