+++ doc/paper/paper.md
@@ -40,7 +40,7 @@ Image registration is required whenever images are taken at different points in 
 Diffeomorphic image registration is an indispensable tool in medical image analysis [@Sotiras:2013a]. Computing diffeomorphisms that map one image to another is expensive. Deformable image registration is an infinite-dimensional problem that upon discretization leads to nonlinear optimality systems with millions or even billions of unknowns. For example, registering two typical medical imaging datasets of size $256^3$ necessitates solving for about 50 million unknowns. Additional complications are the ill-posedness and non-linearty of this inverse problem [@Fischer:2008a]. Consequently, image registration can take several minutes on multi-core high-end CPUs. Many of the available methods reduce the number of unknowns by using coarser resolutions either through parameterization or by solving the problem on coarser grids; they use simplified algorithms and deliver subpar registration quality. In the age of big data, clinical population studies that require thousands of registrations are incresingly common, and execution times of individual registrations become more critical. We provide technology that allows solving registration problems for clinical datasets in seconds. In addition, we have made available to the public a software that works on multi-node, multi-GPU architectures [@Brunn:2020a; @Brunn:2021a] that allows the registration of large-scale microscopic imaging data such as CLARITY imaging [@Tomer:2014a; @Kutten:2017a].
 
 # Highlights
-`CLAIRE` can be used to register images of $2048^3$ (25 B unknowns) on 64 nodes with 256 GPUs on TACC’s Longhorn system [@Brunn:2020a]. `CLAIRE` has been used for the registration of high resolution CLARITY imaging data [@Brunn:2020a]. The GPU version of `CLAIRE` can solve clinically relevant problem sizes (50 M unknowns) in approximately 5 seconds on a single NVIDIA Tesla V100 [@Brunn:2020a]. `CLAIRE` has also been applied to hundreds of images in brain tumor imaging studies [@Mang:2017c; @Scheufele:2021a], and integrated with models for biophysics inversion [@Mang:2018a; @Mang:2020a; @Scheufele:2020a; @Scheufele:2019a; @Scheufele:2021a; @Subramanian:2020b] and Alzheimer's disease progression [@Scheufele:2020c]. `CLAIRE` uses highly optimized computational kernels and effective, state-of-the-art algorithms for time integration and numerical optimization. Our most recent version of `CLAIRE` features a Python interface to assist users in their applications.
+`CLAIRE` can be used to register images of $2048^3$ (25 B unknowns) on 64 nodes with 256 GPUs on TACC’s Longhorn system [@Brunn:2020a]. `CLAIRE` has been used for the registration of high resolution CLARITY imaging data [@Brunn:2020a]. The GPU version of `CLAIRE` can solve clinically relevant problem sizes (50 M unknowns) in approximately 5 seconds on a single NVIDIA Tesla V100 [@Brunn:2020a]. `CLAIRE` has also been applied to hundreds of images in brain tumor imaging studies [@Bakas:2018a; @Mang:2017c; @Scheufele:2021a], and integrated with models for biophysics inversion [@Mang:2018a; @Mang:2020a; @Scheufele:2020a; @Scheufele:2019a; @Scheufele:2021a; @Subramanian:2020b] and Alzheimer's disease progression [@Scheufele:2020c]. `CLAIRE` uses highly optimized computational kernels and effective, state-of-the-art algorithms for time integration and numerical optimization. Our most recent version of `CLAIRE` features a Python interface to assist users in their applications.
 
 We provide a detailed documentation on how to execute, compile, and install `CLAIRE` on various systems at our deployment page <https://andreasmang.github.io/claire>.
 
