### 2021-02-13 #3038 @andreasmang
> Hi @andreasmang - while the reviewers start working through their checklists, can you address the missing DOIs in the references?

@mjsottile: Thanks. I have added DOIs to all articles for which they exist to the bib file. Changes have been pushed to the repo. Let me know if there are any issues...
