### 2021-02-14 4d8c05c
mean subtracted NCC working
