### 2021-02-15 andreasmang/claire#4 @NMontanaBrown
JOSS Review: Unit Tests
# Issue Description

Hi @andreasmang, how you have tested/validated your code and it's outputs? Currently, as far as I can find, there is no automated test suite via Github Actions or other.
