### 2021-02-15 andreasmang/claire#5 @NMontanaBrown
JOSS Review: Contributing Guidelines
# Issue Description

Hi @andreasmang, what are the contributing guidelines for this project? They are missing from the repository.
