### 2021-02-15 #3038 @NMontanaBrown
Hi team, I've started review with two specific issues.

[Unit tests](https://github.com/andreasmang/claire/issues/4)
[Contributing Guidelines](https://github.com/andreasmang/claire/issues/5)
