+++ README.md
@@ -1,6 +1,8 @@
 # CLAIRE
 
 [![status](https://img.shields.io/badge/arXiv-1808.04487-red)](https://arxiv.org/abs/1808.04487)
+[![status](https://img.shields.io/badge/licencse-GPL-blue)](https://github.com/andreasmang/claire/tree/master/LICENSE)
+
 
 * Are you looking for **examples**? Check the [doc/examples](https://github.com/andreasmang/claire/tree/master/doc/examples) folder.
 * Are in interested in **how CLAIRE works**? Check the [documentation](#clairedoc).
