+++ doc/README-REFERENCES.md
@@ -1,4 +1,4 @@
-# References
+# CLAIRE: References
 
 Go back to [README.md](../README.md).
 
