+++ src/CLAIRE.cpp
@@ -72,6 +72,9 @@ PetscErrorCode CLAIRE::Initialize() {
     this->m_AdjointVariable = NULL;     ///< adjoint variable
     this->m_IncStateVariable = NULL;    ///< incremental state variable
     this->m_IncAdjointVariable = NULL;  ///< incremental adjoint variable
+    
+    this->m_CoarseReg = nullptr;
+    this->m_CoarseRegOpt = nullptr;
 
     PetscFunctionReturn(ierr);
 }
@@ -104,6 +107,12 @@ PetscErrorCode CLAIRE::ClearMemory(void) {
     cudaPrintDeviceMemory();
   }
 #endif
+    
+    ierr = Free(this->m_CoarseReg); CHKERRQ(ierr);
+    if (this->m_CoarseRegOpt) {
+      this->m_CoarseRegOpt->m_FFT.fft = nullptr;
+    }
+    ierr = Free(this->m_CoarseRegOpt); CHKERRQ(ierr);
 
     // delete all variables
     ierr = this->ClearVariables(); CHKERRQ(ierr);
@@ -139,6 +148,57 @@ PetscErrorCode CLAIRE::ClearVariables(void) {
     PetscFunctionReturn(ierr);
 }
 
+PetscErrorCode CLAIRE::CreateCoarseReg() {
+  PetscErrorCode ierr = 0;
+  PetscFunctionBegin;
+  
+  ierr = Assert(this->m_CoarseRegOpt != nullptr, "coarse grid RegOpt not initialized"); CHKERRQ(ierr);
+  
+  ierr = AllocateOnce<CLAIRE>(this->m_CoarseReg, this->m_CoarseRegOpt); CHKERRQ(ierr);
+  
+  PetscFunctionReturn(ierr);
+}
+
+PetscErrorCode CLAIRE::InitializeCoarseReg() {
+  PetscErrorCode ierr = 0;
+  PetscFunctionBegin;
+  
+  ierr = Assert(this->m_Opt != nullptr, "fine grid RegOpt not initialized"); CHKERRQ(ierr);
+  
+  //printf("%p\n", this->m_Opt->m_FFT_coarse.fft);
+  //ierr = Assert(this->m_Opt->m_FFT.fft != nullptr, "no FFT"); CHKERRQ(ierr);
+  if (this->m_Opt->m_FFT_coarse.fft) {
+    
+    ierr = AllocateOnce(this->m_CoarseRegOpt, *this->m_Opt); CHKERRQ(ierr);
+  
+    this->m_CoarseRegOpt->m_FFT = this->m_Opt->m_FFT_coarse;
+    this->m_CoarseRegOpt->m_FFT_coarse.fft = nullptr;
+    this->m_CoarseRegOpt->m_Domain.level = this->m_Opt->m_Domain.level;
+    this->m_CoarseRegOpt->m_Domain.nl = 1;
+    this->m_CoarseRegOpt->m_Domain.ng = 1;
+    this->m_CoarseRegOpt->m_Domain.mpicomm = this->m_Opt->m_Domain.mpicomm;
+    this->m_CoarseRegOpt->m_Domain.rowcomm = this->m_Opt->m_Domain.rowcomm;
+    this->m_CoarseRegOpt->m_Domain.colcomm = this->m_Opt->m_Domain.colcomm;
+    for (int i=0; i<3; ++i) {
+      this->m_CoarseRegOpt->m_Domain.nx[i] = this->m_Opt->m_FFT_coarse.nx[i];
+      this->m_CoarseRegOpt->m_Domain.istart[i] = this->m_Opt->m_FFT_coarse.istart[i];
[…]
