+++ src/Solver/TransportEquationSL.cpp
@@ -530,10 +530,10 @@ PetscErrorCode TransportEquationSL::SolveIncForwardProblem() {
         ierr = AllocateOnce(this->m_SemiLagrangianMethod, this->m_Opt); CHKERRQ(ierr);
         ierr = this->m_SemiLagrangianMethod->SetWorkVecField(this->m_WorkVecField[0]); CHKERRQ(ierr);
         ierr = this->m_SemiLagrangianMethod->ComputeTrajectory(this->m_VelocityField, "state"); CHKERRQ(ierr);
-    } else if (this->m_Opt->m_KrylovMethod.pctype == TWOLEVEL) {
+    }/* else if (this->m_Opt->m_KrylovMethod.pctype == TWOLEVEL) {
       ierr = this->m_SemiLagrangianMethod->SetWorkVecField(this->m_WorkVecField[0]); CHKERRQ(ierr);
       ierr = this->m_SemiLagrangianMethod->ComputeTrajectory(this->m_VelocityField, "state"); CHKERRQ(ierr);
-    }
+    }*/
     //ierr = this->m_SemiLagrangianMethod->ComputeTrajectory(this->m_VelocityField, "state"); CHKERRQ(ierr);
 
     if (this->m_Opt->m_OptPara.method == FULLNEWTON) {   // gauss newton
@@ -693,10 +693,10 @@ PetscErrorCode TransportEquationSL::SolveIncAdjointEquationGN() {
         ierr = AllocateOnce(this->m_SemiLagrangianMethod, this->m_Opt); CHKERRQ(ierr);
         ierr = this->m_SemiLagrangianMethod->SetWorkVecField(this->m_WorkVecField[0]); CHKERRQ(ierr);
         ierr = this->m_SemiLagrangianMethod->ComputeTrajectory(this->m_VelocityField, "adjoint"); CHKERRQ(ierr);
-    } else if (this->m_Opt->m_Domain.level > 0) {
+    }/* else if (this->m_Opt->m_Domain.level > 0) {
       ierr = this->m_SemiLagrangianMethod->SetWorkVecField(this->m_WorkVecField[0]); CHKERRQ(ierr);
       ierr = this->m_SemiLagrangianMethod->ComputeTrajectory(this->m_VelocityField, "adjoint"); CHKERRQ(ierr);
-    }
+    }*/
 
     // compute divergence of velocity field
     //ierr = this->m_WorkScaField[1]->GetArrayWrite(kernel.pDivV); CHKERRQ(ierr);
