+++ src/Preconditioner.cpp
@@ -210,6 +210,7 @@ PetscErrorCode Preconditioner::ClearMemory() {
     if (this->m_CoarseGrid->m_Opt != NULL) {
         this->m_CoarseGrid->m_Opt->ProcessTimers();
         this->m_CoarseGrid->m_Opt->WriteLogFile(true);
+        this->m_CoarseGrid->m_Opt->m_FFT.fft = nullptr;
         delete this->m_CoarseGrid->m_Opt;
         this->m_CoarseGrid->m_Opt = NULL;
     }
@@ -401,14 +402,33 @@ PetscErrorCode Preconditioner::SetupCoarseGrid() {
         ierr = reg::ThrowError("allocation failed"); CHKERRQ(ierr);
     }
     this->m_CoarseGrid->m_Opt->m_Domain.level = this->m_Opt->m_Domain.level + 1;
+    this->m_CoarseGrid->m_Opt->m_FFT = this->m_Opt->m_FFT_coarse;
+    this->m_CoarseGrid->m_Opt->m_FFT_coarse.fft = nullptr;
+    this->m_CoarseGrid->m_Opt->m_Domain.nl = 1;
+    this->m_CoarseGrid->m_Opt->m_Domain.ng = 1;
+    this->m_CoarseGrid->m_Opt->m_Domain.mpicomm = this->m_Opt->m_Domain.mpicomm;
+    this->m_CoarseGrid->m_Opt->m_Domain.rowcomm = this->m_Opt->m_Domain.rowcomm;
+    this->m_CoarseGrid->m_Opt->m_Domain.colcomm = this->m_Opt->m_Domain.colcomm;
+    for (int i=0; i<3; ++i) {
+      this->m_CoarseGrid->m_Opt->m_Domain.nx[i] = this->m_Opt->m_FFT_coarse.nx[i];
+      this->m_CoarseGrid->m_Opt->m_Domain.istart[i] = this->m_Opt->m_FFT_coarse.istart[i];
+      this->m_CoarseGrid->m_Opt->m_Domain.isize[i] = this->m_Opt->m_FFT_coarse.isize[i];
+      this->m_CoarseGrid->m_Opt->m_Domain.ng *= this->m_CoarseGrid->m_Opt->m_Domain.nx[i];
+      this->m_CoarseGrid->m_Opt->m_Domain.nl *= this->m_CoarseGrid->m_Opt->m_Domain.isize[i];
+      this->m_CoarseGrid->m_Opt->m_Domain.hx[i] = 2.*PETSC_PI/static_cast<ScalarType>(this->m_CoarseGrid->m_Opt->m_Domain.nx[i]);
+    }
+    this->m_CoarseGrid->m_Opt->m_SetupDone = true;
+    //this->m_CoarseGrid->m_Opt->m_KrylovMethod.matvectype = PRECONDMATVECSYM;
+    //this->m_CoarseGrid->m_Opt->m_KrylovMethod.pctype = NOPC;
 
     // get grid scale and compute number of grid points
-    scale = this->m_Opt->m_KrylovMethod.pcgridscale;
+    /*scale = this->m_Opt->m_KrylovMethod.pcgridscale;
     for (int i = 0; i < 3; ++i) {
         value = static_cast<ScalarType>(this->m_Opt->m_Domain.nx[i])/scale;
         this->m_CoarseGrid->m_Opt->m_Domain.nx[i] = static_cast<IntType>(std::ceil(value));
     }
-    ierr = this->m_CoarseGrid->m_Opt->DoSetup(false); CHKERRQ(ierr);
+    ierr = this->m_CoarseGrid->m_Opt->DoSetup(false); CHKERRQ(ierr);*/
+    
[…]
