+++ include/interp3_gpu_new.hpp
@@ -63,7 +63,7 @@ void checkDomain(short* which_proc, ScalarType* xq, ScalarType* yq, ScalarType* 
 
 void printGPU3DVector(ScalarType* arr1, ScalarType* arr2, ScalarType* arr3, int nq);
 
-void initializeGrid(ScalarType* xq, ScalarType* yq, ScalarType* zq, ScalarType* f, ScalarType* ref, ScalarType* h, int* isize, int* istart, int* nx, int caseid);
+void initializeGrid(ScalarType* xq, ScalarType* yq, ScalarType* zq, ScalarType* f, ScalarType* ref, ScalarType* h, IntType* isize, IntType* istart, IntType* nx, int caseid);
 
 void test(ScalarType* f, int nq);
 
