### 2021-03-02 #3038 @tjasaki
Yes @andreasmang.  I did not mean to imply that the fault could be all at your end.  For example, I needed to install cmake on my system.  Later today I can pass along more particulars about how the build went.  We'll get there.  Tom
