### 2021-03-02 #3038 @andreasmang
@tjasaki Sorry. I also did not want to imply anything. Feedback is very much appreciated, especially when it comes to compilation related issues. If you want me to update the documentation considering particular issues, do not hesitate. We have several folks involved building this on different systems but they are mostly local (as in I work with them directly) and it's quite complicated due to the different dependencies and the fact that it is developed for HPC architectures...Plan is to increase the visibility and get more folks to use the software. So, again, I am grateful for any hints that help us iron out issues.
