### 2021-03-02 #3038 @andreasmang
Also, we are currently pushing everything to the GPU. There are less dependencies on that branch, since the FFT kernels are now part of CLAIRE...This will become the default branch soon...
