### 2021-03-05 61a1d43
fixed GPU kernels for 2level use
