+++ src/TaoInterface.cpp
@@ -46,7 +46,7 @@ PetscErrorCode EvaluateObjective(Tao tao, Vec x, ScalarType* Jx, void* ptr) {
     (void)tao;
     optprob = reinterpret_cast<OptimizationProblem*>(ptr);
     ierr = Assert(optprob != NULL, "null pointer"); CHKERRQ(ierr);
-
+    
     // compute objective value
     ierr = optprob->EvaluateObjective(Jx, x); CHKERRQ(ierr);
 
@@ -555,9 +555,11 @@ PetscErrorCode OptimizationMonitor(Tao tao, void* ptr) {
 
     // save gradient norm
     optprob->GetOptions()->m_Monitor.gradnorm = gnorm;
-
+    
+    if (iter > 0) {
     // remember current iterate
-    optprob->IncrementIterations();
+      optprob->IncrementIterations();
+    }
 
     // tao: display convergence reason
     if (optprob->GetOptions()->m_Verbosity > 0) {
