+++ src/Differentiation/DifferentiationFD.cpp
@@ -105,20 +105,23 @@ PetscErrorCode DifferentiationFD::SetupData(ScalarType *x1, ScalarType *x2, Scal
     }*/
     
     const IntType nghost = 4;
-    const IntType halo[3] = {nghost, 0, 0};
+    halo[0] = nghost; halo[1] = 0; halo[2] = 0;
 
 #if defined(REG_HAS_CUDA) || defined(REG_HAS_MPICUDA)
     if (this->m_Opt->rank_cnt == 1) {
+      isizeg[0] = this->m_Opt->m_Domain.isize[0];
+      isizeg[1] = this->m_Opt->m_Domain.isize[1];
+      isizeg[2] = this->m_Opt->m_Domain.isize[2];
       this->mtex = gpuInitEmptyGradientTexture(this->m_Opt->m_Domain.nx);
-      ierr = initConstants(this->m_Opt->m_Domain.isize, 
+      /*ierr = initConstants(this->m_Opt->m_Domain.isize, 
                            this->m_Opt->m_Domain.isize,
                            this->m_Opt->m_Domain.hx,
-                           halo); CHKERRQ(ierr);
+                           halo); CHKERRQ(ierr);*/
     } else {
       ierr = AllocateOnce(this->m_GhostPlan, this->m_Opt, nghost); CHKERRQ(ierr);
-      IntType isize_g[3], istart_g[3];
+      IntType istart_g[3];
       //size_t nlghost;
-      size_t g_alloc_max = this->m_GhostPlan->get_ghost_local_size_x(isize_g, istart_g);
+      size_t g_alloc_max = this->m_GhostPlan->get_ghost_local_size_x(isizeg, istart_g);
       //nlghost = isize_g[0]*isize_g[1]*isize_g[2];
 
       // ghost data mem alloc on CPU
@@ -129,11 +132,11 @@ PetscErrorCode DifferentiationFD::SetupData(ScalarType *x1, ScalarType *x2, Scal
       // ghost data mem alloc on GPU
       //cudaMalloc((void**)&this->d_Ghost, this->nlghost*sizeof(ScalarType));
       
-      this->mtex = gpuInitEmptyGradientTexture(isize_g); CHKERRQ(ierr);
-      ierr = initConstants(this->m_Opt->m_Domain.isize,
+      this->mtex = gpuInitEmptyGradientTexture(isizeg); CHKERRQ(ierr);
+      /*ierr = initConstants(this->m_Opt->m_Domain.isize,
                            isize_g,
                            this->m_Opt->m_Domain.hx,
-                           halo); CHKERRQ(ierr);
+                           halo); CHKERRQ(ierr);*/
     }
 #else
     ierr = DebugNotImplemented(); CHKERRQ(ierr);
@@ -213,9 +216,9 @@ PetscErrorCode DifferentiationFD::Gradient(ScalarType *g1,
       ZeitGeist_tick(FD_COMM);
       this->m_GhostPlan->share_ghost_x(m, this->m_Ghost); 
       ZeitGeist_tock(FD_COMM);
-      ierr = computeGradient(g1, g2, g3, this->m_Ghost, this->mtex, this->m_Opt->m_Domain.isize, this->m_Opt->m_Domain.hx, true); CHKERRQ(ierr);
[…]
