### 2021-03-05 fb1a89b
C++11 constexpr fix
