### 2021-03-09 dfe6674
code of conduct and contributing guidlines added
