+++ doc/README-INSTALL.md
@@ -1,22 +1,84 @@
 # CLAIRE: Detailed Installation Guide
 
+Go back to [README.md](../README.md).
 
 ## Content
 
-* [Requirements](#requirements)
-* [Dependencies](#dependencies)
-	* Required Dependencies
-  * Step 1: Downloading Dependencies
-  * Step 2: Installing Dependencies
-  	* Quick Shot
-  	* Detailed Instructions
-  * Step 3: Setting Environment Variables
-* [Building CLAIRE](#buildclaire)
+* [Installation Overview](#installation)
+* [Detailed Installation Guide](#verboseinstall)
+	* [Requirements](#requirements)
+	* [Dependencies](#dependencies)
+		* Required Dependencies and Compatibility
+		* Step 1: Downloading and Installing Dependencies
+		* Step 2: Setting Environment Variables
+	* [Building CLAIRE](#buildclaire)
 * [Additional Info for Dependencies](#depsinf)
-* [Troubleshooting](#faq)
+* [Troubleshooting and Known Issues](#faq)
 
 
-## Requirements <a name="requirements"></a>
+
+# CLAIRE: Quick Installation Guide<a name="installation"></a>
+
+Here, we only provide a minimal installation guide. We provide a make environment to download and install the dependencies using generic settings that have worked on our systems. If this brief installation guide does not work for you, please consult the [detailed installation guide](#verbosiinstall) below.
+
+## Contents
+1. [One Shot](#oneshot)
+2. [Step by Step](#stepbystep)
+
+## One Shot <a name="oneshot"></a>
+
+```bash
+cd deps
+make
+source env_source.sh
+cd ..
+make -j
+./bin/claire -synthetic 0
+```
+
+The enviroment variable needs to be sourced every time you log out of your computer or start a new bash. As an alternative, you can add the content of `env_source.sh` to your `.bashrc` or `bash_profile`.
+
+## Step by Step  <a name="stepbystep"></a>
+
+### Step 1) Installing Dependencies
+
+To install the dependencies go to the top level directory of CLAIRE in your command window and execute the following commands within your command window:
+
+```bash
+cd deps
+make
+```
+
+This makefile downloads and compiles the dependencies for CLAIRE. To add these dependencies to your environment type the following into your command line and press return:
+
+```bash
+source env_source.sh
+```
+
+The enviroment variable needs to be sourced every time you log out of your computer or start a new bash. As an alternative, you can add the content of `env_source.sh` to your `.bashrc` or `bash_profile`.
+
+### Step 2) Compiling CLAIRE
+
+Assuming that your in the top level directory of CLAIRE, all you need to do is to type
+
+```bash
+make -j
+```
+
+### Step 3) Executing CLAIRE
+
+If you would like to verify if CLAIRE has been installed correctly run the following command in your command window:
+
+```bash
+./bin/claire -synthetic 0
+```
+
+Additional examples for executing CLAIRE are described in [doc/README-RUNME.md](doc/README-RUNME.md)
+
+
+## Detailed Installation Guide <a name="verboseinstall"></a>
+
+### Requirements <a name="requirements"></a>
 
 * MPI (Open MPI; MVAPICH; Intel MPI); required by `PETSc`, and `CLAIRE`
 * cmake ([https://cmake.org](https://cmake.org)); required by `niftilib`
@@ -24,19 +86,35 @@
[…]
