+++ doc/paper/paper.md
@@ -1,48 +1,67 @@
 ---
-title: 'CLAIRE: Constrained Large Deformation Diffeomorphic Image Registration on HPC platforms'
+title: 'CLAIRE: Constrained Large Deformation Diffeomorphic Image Registration on Parallel Computing Architectures'
 tags:
-  - c++
+  - C++
+  - GPUs
+  - parallel computing
+  - high performance computing
   - large deformation diffeomorphic registration
   - optimal control
   - medical imaging
 authors:
   - name: Malte Brunn
-    affiliation: 1
+    affiliation: "1"
   - name: Naveen Himthani
-    affiliation: 2
+    affiliation: "2"
   - name: George Biros
-    affiliation: 2
+    affiliation: "2"
   - name: Miriam Mehl
-    affiliation: 1
-  - name: Andreas Mang
-    affiliation: 3
+    affiliation: "1"
+  - name: Andreas Mang^[Corresponding Author]
+    orcid: 0000-0003-4605-3290
+    affiliation: "3"
 affiliations:
- - name: Computer Science, University Stuttgart
+ - name: Department of Computer Science, University Stuttgart
    index: 1
- - name: Oden Institute, University of Texas at Austin
+ - name: Oden Institute for Computational Engineering and Sciences, University of Texas at Austin
    index: 2
- - name: Mathematics, University of Houston
+ - name: Department of Mathematics, University of Houston
    index: 3
-date: 25 August 2020
+date: 21 Januar 2021
 bibliography: paper.bib
 ---
 
 # Summary
-
+[`CLAIRE`](https://andreasmang.github.io/claire) [@claire-web] is a computational framework for **C**onstrained **LA**rge deformation diffeomorphic **I**mage **RE**gistration [@Mang:2019a]. It supports highly-optimized, parallel computational kernels for (multi-node) CPU [@Mang:2016a; @Gholami:2017a; @Mang:2019a] and (multi-node multi-)GPU architectures [@Brunn:2020a; @Brunn:2021a]. `CLAIRE` uses MPI for distributed-memory parallelism and can be scaled up to thousands of cores [@Mang:2019a; @Mang:2016a] and GPU devices [@Brunn:2020a]. The multi-GPU implementation exploiting device direct communication. The computational kernels are interpolation for semi-Lagrangian time integration, and a mixture of high-order finite difference operators and Fast-Fourier-Transforms (FFTs) for differentiation. `CLAIRE` uses a Newton--Krylov solver for numerical optimization [@Mang:2015a; @Mang:2017a]. It features various schemes for regularization of the control problem [@Mang:2016a] and different similarity measures. `CLAIRE` features different preconditioners for the reduced space Hessian [@Brunn:2020a; @Mang:2019a] to optimize computational throughput and enable fast convergence. It uses `PETSc` [@petsc-web] for scalable and efficient linear algebra operations and solvers and `TAO` [@petsc-web; @Munson:2015a] for numerical optimization. `CLAIRE` can be downloaded at <https://github.com/andreasmang/claire>.
 
 # Statement of Need
[…]
