+++ doc/paper/paper.bib
@@ -1,95 +1,311 @@
+%% This BibTeX bibliography file was created using BibDesk.
+%% https://bibdesk.sourceforge.io/
+
+%% Created for Andreas Mang at 2021-02-13 16:17:55 -0600 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
+@article{Bakas:2018a,
+	author = {S. Bakas and M. Reyes and A. Jakab and S. Bauer and M. Rempfler and others},
+	date-added = {2021-02-13 16:16:34 -0600},
+	date-modified = {2021-02-13 16:17:55 -0600},
+	journal = {arXiv},
+	number = {1811.02629},
+	title = {Identifying the best machine learning algorithms for brain tumor segmentation, progression assessment, and overall survival prediction in the {BRATS} challenge},
+	year = {2018}}
+
+@inproceedings{Kutten:2017a,
+	author = {Kutten, K. S. and Charon, N. and Miller, M. I. and Ratnanather, J. T. and Deisseroth, K. and Ye, L. and Vogelstein, J. T.},
+	booktitle = {Proc Medical Image Computing and Computer-Assisted Intervention},
+	date-added = {2020-04-18 23:23:16 -0500},
+	date-modified = {2021-02-13 15:50:55 -0600},
+	doi = {10.1007/978-3-319-66182-7_32},
+	pages = {275--282},
+	title = {A Large Deformation Diffeomorphic Approach to Registration of {CLARITY} Images via Mutual Information},
+	volume = {LNCS 10433},
+	year = {2017}}
+
+@article{Tomer:2014a,
+	author = {Tomer, Raju and Ye, Li and Hsueh, Brian and Deisseroth, Karl},
+	date-added = {2017-03-28 02:58:57 +0000},
+	date-modified = {2021-02-13 15:46:06 -0600},
+	doi = {10.1038/nprot.2014.123},
+	journal = {Nature protocols},
+	number = {7},
+	pages = {1682--1697},
+	title = {Advanced {CLARITY} for rapid and high-resolution imaging of intact tissues},
+	volume = {9},
+	year = {2014}}
+
+@article{Trouve:1998a,
+	author = {Trouv\'e, A.},
+	date-added = {2020-11-04 16:50:33 -0600},
+	date-modified = {2021-02-13 15:39:43 -0600},
+	doi = {10.1023/A:1008001603737},
+	journal = {International Journal of Computer Vision},
+	number = {3},
+	pages = {213--221},
+	title = {Diffeomorphism groups and pattern matching in image analysis},
+	volume = {28},
+	year = {1998}}
+
+@book{Younes:2010a,
+	author = {Younes, L.},
+	date-added = {2020-11-04 16:47:32 -0600},
+	date-modified = {2021-02-13 15:44:18 -0600},
+	doi = {10.1007/978-3-642-12055-8},
+	publisher = {Springer-Verlag Berlin Heidelberg},
+	title = {Shapes and diffeomorphisms},
+	year = {2010}}
+
+@article{Beg:2005a,
+	author = {Beg, M.~F. and Miller, M.~I. and A.~Trouv\'e and L.~Younes},
+	date-added = {2020-11-04 16:47:13 -0600},
+	date-modified = {2021-02-13 15:40:48 -0600},
+	doi = {10.1023/B:VISI.0000043755.93987.aa},
+	journal = {International Journal of Computer Vision},
+	number = {2},
+	pages = {139--157},
+	title = {Computing large deformation metric mappings via geodesic flows of diffeomorphisms},
+	volume = {61},
+	year = {2005}}
+
+@inproceedings{Mang:2017c,
[…]
