+++ doc/README-RUNME.md
@@ -6,19 +6,16 @@ Go back to [README.md](../README.md).
 * [Overview](#clairebins)
 * [Get Help](#clairehelp)
 * [Simple Examples: `claire`](#claireexmp)
-  * [Synthetic Problem](#clairexmp1)
-  * [Synthetic Problem (Parallel Execution)](#clairexmp2)
-  * [Real Data (Parallel Execution)](#clairexmp3)
-  * [Regularization Parameter Estimation](#clairexmp4)
-  * [Parameter Continuation](#clairexmp5)
-  * [Output Velocities](#clairexmp6)
+	* [Synthetic Problem](#clairexmp1)
+	* [Synthetic Problem (Parallel Execution)](#clairexmp2)
+	* [Real Data (Parallel Execution)](#clairexmp3)
+	* [Regularization Parameter Estimation](#clairexmp4)
+	* [Parameter Continuation](#clairexmp5)
+	* [Output Velocities](#clairexmp6)
 * [Simple Examples: `clairetools`](#toolsxmp)
-  * [Transporting Images](#toolsxmp1)
-  * [Computing Jacobians](#toolsxmp2)
-* [Testing](#testing)
-	* [Unit Tests](#unittest)
-	* [Numerics](#numerics)
-	* [NIREP Data](#nirep)
+	* [Transporting Images](#toolsxmp1)
+	* [Computing Jacobians](#toolsxmp2)
+* [Testing and Benchmarks](#testing)
 
 
 ## Overview <a name="clairebins"></a>
@@ -154,45 +151,27 @@ mpirun -np 20 $bindir/clairetools -v1 velocity-field-x1.nii.gz       \
 
 
 
-## Testing <a name="testing"></a>
+## Testing and Benchmarks <a name="testing"></a>
 
-We have implemented [unit tests](#unittests) for the main computational kernels available in CLAIRE, tests to study the performance and accuracy of the mathematical operators (see [numerics](#numerics)), and provide a repo for the test data we have considered in our most recent publications [nirep](#nirep).
+We have implemented numerical tests for the main computational kernels available in CLAIRE to study the performance and accuracy of the mathematical operators that appear in the optimality system. For reproducability, we also posted the NIREP data at [https://github.com/andreasmang/nirep](https://github.com/andreasmang/nirep). We have used this data extensively in our most recent [prior work](README-REFERENCES.md).
 
+To build binaries for testing set `BUILD_TEST=yes` in the makefile. This will build two binaries: `benchmark` and `test`. 
 
-### Unit Tests <a name="unittests"></a>
+The `test` application allows users to check the main computational kernels:
+* the interpolation kernels (`-interp` flag; checks interpolation accuracy)
+* the regularization operators (e.g., biharmonic or laplacian regularization operators; `-reg` flag)
+* numerical differentiation (`-diff`; checks differentiation accuracy)
 
-To build binaries for testing set `BUILD_TEST=yes` in the makefile. This will build two binaries: `benchmark` (for numerical tests) and `test` for unit tests. Considering the unit tests implemented in the `test` binary, we provide methods to assess the main computational kernels implemented in CLAIRE, namely:
-* the interpolation kernels (`-interp` flag) 
-* the regularization operators (e.g., biharmonic operators; `-reg` flag)
-* numerical differentiation (`-diff`)
+These tests are implemented in the `*.cpp` files in the [UnitTests](https://github.com/andreasmang/claire/tree/gpu/src/UnitTests) subfolder.
 
-The tests implemented in `benchmark` are described in the [numerics](#numerics) section below.
-
-### Numerics <a name="numerics"></a>
-
-We have implemented several high-level numerical checks to assess the performance of our methodology and ensure that the mathematical operators are correct.
-
-
[…]
