### 2021-03-09 1b45f2b
references updated; installation guide updated
