### 2021-03-17 8b72cd0
typos fixed in installation guide
