+++ doc/README-INSTALL.md
@@ -21,9 +21,11 @@ Go back to [README.md](../README.md).
 
 ## Installation Overview (Quick Guide)<a name="installation"></a>
 
-Here, we only provide a minimal installation guide. We provide a make environment to download and install the dependencies using generic settings that have worked on our systems. If this brief installation guide does not work for you, please consult the [detailed installation guide](#verboseinstall) below.
+In this section, we provide a minimal installation guide. We provide a make environment to download and install the dependencies using generic settings that have worked on most of our systems. If this brief installation guide does not work for you, please consult the [detailed installation guide](#verboseinstall) below.
 
 ### One Shot <a name="oneshot"></a>
+
+To use the default settings to build dependencies and CLAIRE itself do the following:
 
 ```bash
 cd deps
@@ -34,13 +36,15 @@ make -j
 ./bin/claire -synthetic 0
 ```
 
-The enviroment variable needs to be sourced every time you log out of your computer or start a new bash. As an alternative, you can add the content of `env_source.sh` to your `.bashrc` or `bash_profile`.
+The enviroment variables need to be sourced every time you log out of your computer or start a new bash (`source env_source.sh`). As an alternative, you can add the content of `env_source.sh` (for example) to your `.bashrc` or `.bash_profile`.
 
 ### Step by Step  <a name="stepbystep"></a>
 
+Next, we go over the steps outlined above step by step. Again, more details are provided [below](#verboseinstall).
+
 #### Step 1) Installing Dependencies
 
-To install the dependencies go to the top level directory of CLAIRE in your command window and execute the following commands within your command window:
+To install the dependencies (the PETSc and NIFTI libraries) go to the top level directory of CLAIRE in your command window and execute the following commands within your command window:
 
 ```bash
 cd deps
@@ -53,11 +57,12 @@ This makefile downloads and compiles the dependencies for CLAIRE. To add these d
 source env_source.sh
 ```
 
-The enviroment variable needs to be sourced every time you log out of your computer or start a new bash. As an alternative, you can add the content of `env_source.sh` to your `.bashrc` or `bash_profile`.
+Notice that the enviroment variables need to be sourced every time you log out of your computer or start a new bash. As an alternative, you can add the content of `env_source.sh` to your `.bashrc` or `bash_profile`.
+
 
 #### Step 2) Compiling CLAIRE
 
-Assuming that your in the top level directory of CLAIRE, all you need to do is to type
+Assuming that you are in the top level directory of CLAIRE, all you need to do is to type
 
 ```bash
 make -j
@@ -71,35 +76,12 @@ If you would like to verify if CLAIRE has been installed correctly run the follo
 ./bin/claire -synthetic 0
 ```
 
-Additional examples for executing CLAIRE are described in [doc/README-RUNME.md](README-RUNME.md)
+Additional examples for executing CLAIRE are described in [doc/README-RUNME.md](README-RUNME.md).
 
 
 ## Detailed Installation Guide <a name="verboseinstall"></a>
 
-### Requirements <a name="requirements"></a>
-
-* MPI (Open MPI; MVAPICH; Intel MPI); required by `PETSc`, and `CLAIRE`
-* cmake ([https://cmake.org](https://cmake.org)); required by `niftilib`
-* python ([https://www.python.org](https://www.python.org)); required by `PETSc` and the optional `pyclaire` bindings
-* zlib ([https://www.zlib.net/](https://www.zlib.net/)); requiered by `niftilib`
-* CUDA-API
-
[…]
