### 2021-03-17 e5d7bd6
typos fixed in installation guide
