+++ doc/README-RUNME.md
@@ -57,17 +57,17 @@ CLAIRE is a software for 3D diffeomorphic image registration. Supported input fo
 
 In [runclaire01.sh](https://github.com/andreasmang/claire/tree/gpu/doc/examples/runclaire01.sh) we execute CLAIRE for a synthetic test problem of size 32x32x32. We use default settings for our solver:
 ```bash
-$bindir/claire -synthetic 0
+$BINDIR/claire -synthetic 0
 ```
 
-`$bindir` is the directory in which the binary is located. The flag `-synthetic` allows one to select several smooth test problems. To change the problem size simply add the `-nx <n1xn2xn3>` flag, where `<n$i$>` represents the problem size in each spatial direction (i.e., `-nx 128x128x128` executes CLAIRE with a problem size of `128x128x128`.) We recommend executing CLAIRE in parallel for larger problem sizes (see [example 2](#clairexmp2))
+`$BINDIR` is the directory in which the binary is located. The flag `-synthetic` allows one to select several smooth test problems. To change the problem size simply add the `-nx <n1xn2xn3>` flag, where `<n$i$>` represents the problem size in each spatial direction (i.e., `-nx 128x128x128` executes CLAIRE with a problem size of `128x128x128`.) We recommend executing CLAIRE in parallel for larger problem sizes (see [example 2](#clairexmp2))
 
 ### Example 02: Synthetic Problem (Parallel Execution) <a name="clairexmp2"></a>
 
 In [runclaire02.sh](examples/runclaire02.sh) we execute CLAIRE for a synthetic test problem of size 128x128x128 in parallel. We use 20 MPI tasks. We use default settings for our solver:
 
 ```bash
-mpirun -np 20 $bindir/claire -synthetic 0 -nx 128
+mpirun -np 20 $BINDIR/claire -synthetic 0 -nx 128
 ```
 
 The options used with `claire` are explained in [example 1](#clairexmp1). The key difference is the instruction `mpirun -np 20` infront of the executable. This instructs your compute node to use 20 MPI tasks. CLAIRE will determine the processor layout for you. We recommend executing CLAIRE in parallel. If you use `mpiexec` replace `mpirun -np 20` with `mpiexec -n 20`.
@@ -78,7 +78,7 @@ The options used with `claire` are explained in [example 1](#clairexmp1). The ke
 In [runclaire03.sh](https://github.com/andreasmang/claire/tree/gpu/doc/examples/runclaire03.sh) we execute CLAIRE for real medical images (in NIfTI format) of size 128x150x128. We use 20 MPI tasks. The data can be found in the [docs/data](data) subdirectory. We use default settings for our solver:
 
 ```bash
-mpirun -np 20 $bindir/claire -mr $datdir/brain01.nii.gz \
+mpirun -np 20 $BINDIR/claire -mr $datdir/brain01.nii.gz \
                              -mt $datdir/brain02.nii.gz
 ```
 
@@ -92,7 +92,7 @@ mpirun -np 20 $bindir/claire -mr $datdir/brain01.nii.gz \
 In [runclaire04.sh](https://github.com/andreasmang/claire/tree/gpu/doc/examples/runclaire04.sh) we execute CLAIRE to automatically identify an adequate regularization parameter for a given set of images. We use default settings for our solver:
 
 ```bash
-mpirun -np 20 $bindir/claire -mr $datdir/brain01.nii.gz \
+mpirun -np 20 $BINDIR/claire -mr $datdir/brain01.nii.gz \
                              -mt $datdir/brain02.nii.gz -train binary
 ```
 
@@ -104,7 +104,7 @@ Running `claire` on real image data is explained in [example 3](#clairexmp3). We
[…]
