### 2021-03-18 andreasmang/claire#5 @andreasmang
Thanks for pointing this out. We have added contributing guidelines to the repo. The can be found here:

https://github.com/andreasmang/claire/blob/gpu/doc/CONTRIBUTING.md
