### 2021-03-18 andreasmang/claire#4 @andreasmang
Thanks for pointing this out. We have implemented binaries to test CLAIRE. We describe them in more detail in the README (doc/README-RUNME.md). GitHub actions are not yet available for our project. We will keep extending our testing environment and consider automated test suites (such as GitHub Actions) in future releases.
