### 2021-03-18 #3038 @andreasmang
Dear all:

I apologize for the silence. The last couple of weeks have been pretty intense. I will be able to respond more quickly to further requests from now on.

> Comment R1: Hi @andreasmang, how you have tested/validated your code and its outputs? Currently, as far as I can find, there is no automated test suite via Github Actions or other.

We did not discuss this in the README. We currently do not have support for Github Actions. This remains subject to future work.
CLAIRE features a test binary used in our group to make sure the computational kernels produce accurate results. We have added a description of these to the README file:
https://github.com/andreasmang/claire/blob/gpu/doc/README-RUNME.md#testing

We also provide a brief description of how to use CLAIRE and include a link to data we have extensively considered in our prior work:

https://github.com/andreasmang/nirep

I am not going to mention additional details here but rather refer to the "overhauled" README environment.

> Comment R2: Hi @andreasmang, what are the contributing guidelines for this project? They are missing from the repository.

Response: Thanks for pointing this out. These did not exist. We have added contributing guidelines to the repository:
https://github.com/andreasmang/claire/blob/master/doc/CONTRIBUTING.md

We have also included a "Code of Conduct"
https://github.com/andreasmang/claire/blob/gpu/doc/CODE_OF_CONDUCT.md

Based on a prior comment posted above we have also modified the structure of the README.

We have also enabled "GitHub Discussions" to allow the community to make comments and suggestions.

Cheers A
