### 2021-03-18 50e66de
fixed compile warnings
