+++ src/RegToolsOpt.cpp
@@ -24,6 +24,7 @@
 #include <algorithm>
 
 #define _TO_STR(s) #s
+#define TO_STR(s) _TO_STR(s)
 
 // trim from start
 static inline std::string &ltrim(std::string &s) {
@@ -642,7 +643,7 @@ PetscErrorCode RegToolsOpt::DisplayOptions() {
         std::cout << line << std::endl;
         std::time_t result = std::time(NULL);
 #ifdef GIT_VERSION
-        std::cout << " Version " << _TO_STR(GIT_VERSION) << " " << std::asctime(std::localtime(&result));
+        std::cout << " Version " << TO_STR(GIT_VERSION) << " " << std::asctime(std::localtime(&result));
 #else
         std::cout << " " << std::asctime(std::localtime(&result));
 #endif
