### 2021-04-02 #3038 @andreasmang
> @mjsottile : I've been reading and reviewing the new material. I am still stymied by getting claire to install properly. I've made progress, it seems to be down to a CUDA (compatibility?) issue which I haven't been able to sort out yet. The installation instructions seem reasonable and have been notably helpful, but accounting for all possibilities would be unrealistic.

@tjasaki (and others): Let me know if there is anything we should do on our end...(i.e., if you would like us to check if we can emulate a particular setting and if we can figure it out). Feel free to raise an ISSUE or send me an email.
