### 2021-04-05 #3038 @mjsottile
@andreasmang @tjasaki I’d like to not get blocked on install issues if it boils down to CUDA related problems.  The difficulty in installing tools based on CUDA is often more an issue with how NVidia chooses to distribute their software/drivers and not a reflection of the software that uses it.  I would recommend @tjasaki communicate directly with the CLAIRE developers via GitHub issues or email with the specific problems being encountered on installation, and if the issue can’t be resolved, skip the installation part of the review for now.  I don’t think it’s useful if we get bogged down in third party GPU driver/framework problems that are orthogonal to the package under review.  If everything else looks reasonable on review, we can revisit whether more needs to be done related to installation at the end of the refereeing process.
