### 2021-04-15 4203318
notes added for compilation in single precision (know issues)
