+++ doc/README-RUNME.md
@@ -65,7 +65,7 @@ $BINDIR/claire -synthetic 0
 In [runclaire02.sh](examples/runclaire02.sh) we execute CLAIRE for a synthetic test problem of size 128x128x128 in parallel. We use 20 MPI tasks. We use default settings for our solver:
 
 ```bash
-mpirun -np 20 $BINDIR/claire -synthetic 0 -nx 128
+$BINDIR/claire -synthetic 0 -nx 128
 ```
 
 The options used with `claire` are explained in [example 1](#clairexmp1). The key difference is the instruction `mpirun -np 20` infront of the executable. This instructs your compute node to use 20 MPI tasks. CLAIRE will determine the processor layout for you. We recommend executing CLAIRE in parallel. If you use `mpiexec` replace `mpirun -np 20` with `mpiexec -n 20`.
@@ -76,8 +76,8 @@ The options used with `claire` are explained in [example 1](#clairexmp1). The ke
 In [runclaire03.sh](https://github.com/andreasmang/claire/tree/gpu/doc/examples/runclaire03.sh) we execute CLAIRE for real medical images (in NIfTI format) of size 128x150x128. We use 20 MPI tasks. The data can be found in the [docs/data](data) subdirectory. We use default settings for our solver:
 
 ```bash
-mpirun -np 20 $BINDIR/claire -mr $datdir/brain01.nii.gz \
-                             -mt $datdir/brain02.nii.gz
+$BINDIR/claire -mr $datdir/brain01.nii.gz \
+               -mt $datdir/brain02.nii.gz
 ```
 
 **<span style="color:red">Important:</span>** The images have to be **affinely preregistered** (i.e., have the same grid size).
@@ -90,21 +90,22 @@ mpirun -np 20 $BINDIR/claire -mr $datdir/brain01.nii.gz \
 In [runclaire04.sh](https://github.com/andreasmang/claire/tree/gpu/doc/examples/runclaire04.sh) we execute CLAIRE to automatically identify an adequate regularization parameter for a given set of images. We use default settings for our solver:
 
 ```bash
-mpirun -np 20 $BINDIR/claire -mr $datdir/brain01.nii.gz \
-                             -mt $datdir/brain02.nii.gz -train binary
+$BINDIR/claire -mr $datdir/brain01.nii.gz \
+               -mt $datdir/brain02.nii.gz -train binary
 ```
 
 Running `claire` on real image data is explained in [example 3](#clairexmp3). We use a method based on parameter continuation to identify an adequate regularization parameter. The search uses the determinant of the deformation gradient as a "metric". The user can define a lower bound for the Jacobian via the `-jbound <dbl>` option (the upper bound is `1/<dbl>`). To perform the search specify the `-train <type>` option. There are two strategies implemented: A simple reduction of the regularization parameter until the bound is hit (use `-train reduce`) and a binary search (use `-train binary`). The line break (backslash `\`) is only added for readability.
 
+Notice that if you compile CLAIRE in single precision only H1-type regularization operators function properly (there are numerical accuracy issues for H2- and H3-type regularization operators in single precision). To use higher order regularization operators, CLAIRE needs to be compiled in double precision (slower).
 
 ### Example 05: Parameter Continuation <a name="clairexmp5"></a>
 
 In [runclaire05.sh](https://github.com/andreasmang/claire/tree/gpu/doc/examples/runclaire05.sh) we show how to execute CLAIRE using a parameter continuation scheme with a target regularization parameter for the velocity. We use default settings for our solver:
 
 ```bash
-mpirun -np 20 $BINDIR/claire -mr $datdir/brain01.nii.gz \
-                             -mt $datdir/brain02.nii.gz \
-                             -betacont 7.75e-04
[…]
