### 2021-04-16 51717ca
instructions for compilation on Opuntia added
