### 2021-04-28 0bbaed4
readme updated (data conversion)
