+++ doc/README-RUNME.md
@@ -49,7 +49,7 @@ $BINDIR/clairetools -help
 
 ## Input Data <a name="clairedata"></a>
 
-CLAIRE is a software for 3D diffeomorphic image registration. Supported input formats for CLAIRE are images stored as '*.nii', '*.nii.gz' or '*.hdr' and '*.img(.gz)'.
+CLAIRE is a software for 3D diffeomorphic image registration. Supported input formats for CLAIRE are images stored as '*.nii', '*.nii.gz' or '*.hdr' and '*.img(.gz)'. This is the default data format for inputs and outputs generated by CLAIRE.
 
 
 ## Simple Examples: `claire` <a name="claireexmp"></a>
@@ -170,7 +170,7 @@ $BINDIR/clairetools -ifile $datdir/brain01.nii.gz      \
                     -convert 2nc
 ```
 
-The output file will be located in the same folder as the input file, with the same filename (for example, `$datdir/brain01.nii.gz` becomes `$datdir/brain01.nc`).
+The output file will be located in the same folder as the input file, with the same filename (for example, `$datdir/brain01.nii.gz` becomes `$datdir/brain01.nc`). Notice that you need to compile CLAIRE with nifti and pnetcdf enabled.
 
 
 ## Job Submission on Dedicated Systems <a name="hpc"></a>
@@ -180,6 +180,13 @@ The output file will be located in the same folder as the input file, with the s
 
 More information about TACC's Longhorn system can be found at [https://www.tacc.utexas.edu/systems/longhorn](https://www.tacc.utexas.edu/systems/longhorn).
 
+To execute CLAIRE on a single GPU node, interactively, do the following:
+```bash
+idev -N1
+ibrun $BINDIR/claire -synthetic 0 -nx 128
+```
+
+An example for a script for executing CLAIRE on Longhorn can be found here:
 * multi-GPU (2 GPUs one node): [doc/examples/longhorn_mgpu.slurm](examples/longhorn_mgpu.slurm)
 
 
