+++ doc/README-RUNME.md
@@ -5,6 +5,7 @@ Go back to [README.md](../README.md).
 ## Content
 * [Overview](#clairebins)
 * [Get Help](#clairehelp)
+* [Input Data](#clairedata)
 * [Simple Examples: `claire`](#claireexmp)
   * [Synthetic Problem](#clairexmp1)
   * [Synthetic Problem (Parallel Execution)](#clairexmp2)
@@ -42,6 +43,15 @@ claire -help
 ```bash
 clairetools -help
 ```
+
+
+## Input Data <a name="clairedata"></a>
+
+CLAIRE is a software for 3D diffeomorphic image registration. Supported input formats for CLAIRE are images stored as `*.nii`, `*.nii.gz` or `*.hdr` and `*.img(.gz)`. This is the default data format for inputs and outputs generated by CLAIRE. 
+
+CLAIRE can also read and write files in the `netcdf` format (`*.nc`). The parallel `netcdf` format allows for parallel IO. The format can be controlled using the `-format` flag. The default output is `*.nii.gz`. Notice that `netcdf` can be written and read in parallel. Therefore, it is recommended for large-scale imaging applications (say, grid sizes of and above `1024x1024x1024`). The user needs to compile the `netcdf` library to use this format. See [README-INSTALL.md](README-INSTALL.md) for additional details.
+
+
 
 ## Simple Examples: `claire` <a name="claireexmp"></a>
 
