+++ doc/README-INSTALL.md
@@ -56,17 +56,19 @@ Make sure that the standard *wrappers* for `mpicc` and `mpicxx` are available on
 ### Dependencies <a name="dependencies"></a>
 
 CLAIRE requires the following libraries to be installed on your system:
-* `FFTW` [http://www.fftw.org](http://www.fftw.org)
+* `FFTW` [http://www.fftw.org](http://www.fftw.org) [ *required* ]
 	* file: [fftw-3.3.6-pl2.tar.gz](http://www.fftw.org/fftw-3.3.6-pl2.tar.gz)
 	* description: library for computing FFTs
 	* additional details can be found at [http://www.fftw.org](http://www.fftw.org)
-* `AccFFT` [http://accfft.org](http://accfft.org)
+	* individual compilation: `./build_libs.sh --bfftw`
+* `AccFFT` [http://accfft.org](http://accfft.org) [ *required* ]
 	* file: accfft.tar.gz
 	* source code also available on gitub: [https://github.com/amirgholami/accfft](https://github.com/amirgholami/accfft)
 	* description: library to compute FFT in parallel (requires FFTW)
 	* additional details can be found at [http://www.accfft.org](http://www.accfft.org)
 	* requires `FFTW` to be installed
-* `PETSc` [https://www.mcs.anl.gov/petsc/](https://www.mcs.anl.gov/petsc/) (requires `python 2.7`)
+	* individual compilation: `./build_libs.sh --baccft`
+* `PETSc` [https://www.mcs.anl.gov/petsc/](https://www.mcs.anl.gov/petsc/) (requires `python 2.7`) [ *required* ]
 	* file: [petsc-lite-3.9.1.tar.gz](http://ftp.mcs.anl.gov/pub/petsc/release-snapshots/petsc-lite-3.9.1.tar.gz)
 	* source code also available on bitbucket: [https://bitbucket.org/petsc/petsc](https://bitbucket.org/petsc/petsc)
 	* description: library for numerics, linear algebra, and optimization
@@ -74,14 +76,28 @@ CLAIRE requires the following libraries to be installed on your system:
 		* [petsc-lite-3.8.3.tar.gz](http://ftp.mcs.anl.gov/pub/petsc/release-snapshots/petsc-lite-3.8.3.tar.gz)
 		* [petsc-lite-3.7.6.tar.gz](http://ftp.mcs.anl.gov/pub/petsc/release-snapshots/petsc-lite-3.7.6.tar.gz)
 		* [petsc-lite-3.7.0.tar.gz](http://ftp.mcs.anl.gov/pub/petsc/release-snapshots/petsc-lite-3.7.0.tar.gz)
-* `niftilib` [https://sourceforge.net/projects/niftilib/files/nifticlib/](https://sourceforge.net/projects/niftilib/files/nifticlib/)
+	* individual compilation:
+		* for single precision: `./build_libs.sh --bpetscsgl` (precision is inherited by CLAIRE)
+		* for double precision: `./build_libs.sh --bpetscdbl` (precision is inherited by CLAIRE)
+		* for single precision (debug): `build_libs.sh --bpetscdbgsgl` (precision is inherited by CLAIRE)
+		* for double precision (debug): `build_libs.sh --bpetscdbgdbl` (precision is inherited by CLAIRE)
+* `niftilib` [https://sourceforge.net/projects/niftilib/files/nifticlib/](https://sourceforge.net/projects/niftilib/files/nifticlib) [ *required* ]
 	* file: [nifticlib-2.0.0.tar.gz](https://sourceforge.net/projects/niftilib/files/nifticlib/nifticlib_2_0_0)
 	* description: library to read and write NIFTI images
[…]
