+++ makefile
@@ -1,11 +1,19 @@
-CXX=mpicxx            # select the MPI compiler on your system
-USEINTEL=no           # enable if you want plan on using intel compilers
-USEINTELMPI=no        # enable if you plan on using INTEL MPI
-USESINGLE=yes         # compile CLAIRE in single precision (needs PETSC in single precision)
-USEPNETCDF=no         # use PNETCDF as data format (do not forget to build the dependency)
-USENIFTI=yes          # use NIFTI as data format; this is the default for IO and should always be enabled
-USEHASWELL=no         # build CLAIRE for a HASWELL architecture
-BUILDTOOLS=yes        # build the `clairetools` binary
+# select MPI compiler
+CXX=mpicxx
+# use intel compilder (yes,no) 
+USEINTEL=no
+# use intel MPI (yes,no) 
+USEINTELMPI=no
+# use single precision (yes,no) 
+USESINGLE=yes
+# use PNETCDF data format (yes,no; requires PNETCDF lib; see dependencies) 
+USEPNETCDF=no
+# use NIFTI data format (yes,no; default IO format; see dependencies)
+USENIFTI=yes
+# build CLAIRE for a HASWELL architecture (yes,no)
+USEHASWELL=no
+# build `clairetools` binary (yes,no)
+BUILDTOOLS=yes
 
 include config/setup.mk
 include config/files.mk
