### 2021-05-28 #3038 @mjsottile
@andreasmang I've looked over the submission myself in addition to the review from @tjasaki and I am satisfied that it meets the requirements for JOSS.  The last steps:

- Please create an archive on zenodo or figshare for the submission
- Ensure that the version for the JOSS submission is correct.
- Ensure that the author list in the paper and archive are correct.

Once you've provided that information I can issue the final commands to pass the submission to an EIC for final processing.
