+++ doc/paper/paper.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Andreas Mang at 2021-02-13 16:17:55 -0600 
+%% Created for Andreas Mang at 2021-05-28 10:41:53 -0500 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -26,19 +26,21 @@
 	pages = {275--282},
 	title = {A Large Deformation Diffeomorphic Approach to Registration of {CLARITY} Images via Mutual Information},
 	volume = {LNCS 10433},
-	year = {2017}}
+	year = {2017},
+	Bdsk-Url-1 = {https://doi.org/10.1007/978-3-319-66182-7_32}}
 
 @article{Tomer:2014a,
 	author = {Tomer, Raju and Ye, Li and Hsueh, Brian and Deisseroth, Karl},
 	date-added = {2017-03-28 02:58:57 +0000},
-	date-modified = {2021-02-13 15:46:06 -0600},
+	date-modified = {2021-05-28 10:41:53 -0500},
 	doi = {10.1038/nprot.2014.123},
-	journal = {Nature protocols},
+	journal = {Nature Protocols},
 	number = {7},
 	pages = {1682--1697},
 	title = {Advanced {CLARITY} for rapid and high-resolution imaging of intact tissues},
 	volume = {9},
-	year = {2014}}
+	year = {2014},
+	Bdsk-Url-1 = {https://doi.org/10.1038/nprot.2014.123}}
 
 @article{Trouve:1998a,
 	author = {Trouv\'e, A.},
@@ -50,7 +52,8 @@
 	pages = {213--221},
 	title = {Diffeomorphism groups and pattern matching in image analysis},
 	volume = {28},
-	year = {1998}}
+	year = {1998},
+	Bdsk-Url-1 = {https://doi.org/10.1023/A:1008001603737}}
 
 @book{Younes:2010a,
 	author = {Younes, L.},
@@ -59,7 +62,8 @@
 	doi = {10.1007/978-3-642-12055-8},
 	publisher = {Springer-Verlag Berlin Heidelberg},
 	title = {Shapes and diffeomorphisms},
-	year = {2010}}
+	year = {2010},
+	Bdsk-Url-1 = {https://doi.org/10.1007/978-3-642-12055-8}}
 
 @article{Beg:2005a,
 	author = {Beg, M.~F. and Miller, M.~I. and A.~Trouv\'e and L.~Younes},
@@ -71,7 +75,8 @@
 	pages = {139--157},
 	title = {Computing large deformation metric mappings via geodesic flows of diffeomorphisms},
 	volume = {61},
-	year = {2005}}
+	year = {2005},
+	Bdsk-Url-1 = {https://doi.org/10.1023/B:VISI.0000043755.93987.aa}}
 
 @inproceedings{Mang:2017c,
 	author = {A. Mang and S. Tharakan and A. Gholami and N. Nimthani and S. Subramanian and J. Levitt and M. Azmat and K. Scheufele and M. Mehl and C. Davatzikos and B. Barth and G. Biros},
@@ -90,7 +95,8 @@
 	journal = {Annual Review of Biomedical Engineering},
 	title = {Integrated Biophysical Modeling and Image Analysis: {A}pplication to Neuro-Oncology},
 	volume = {22},
-	year = {2020}}
+	year = {2020},
+	Bdsk-Url-1 = {https://doi.org/10.1146/annurev-bioeng-062117-121105}}
 
 @article{Sotiras:2013a,
 	author = {Sotiras, A. and Davatzikos, C. and Paragios, N.},
@@ -102,7 +108,8 @@
[…]
