### 2021-05-28 #3038 @andreasmang
@mjsottile I have looked carefully over the submission.

- I have fixed the affiliations of my co-authors.
- I have fixed three minor typos in the text and references.
- I have created an archive on zenodo (https://doi.org/10.5281/zenodo.4837427; I am new to this so let me know if I missed anything; it is a mirror of v0.08 of the code repo; i.e., the current version of the GPU branch). The ZIP file has the version number and ID in the filename.
- I have pushed a tag to the repo (v0.08) for the JOSS submission (the current version of the code):

Let me know if further action is required on my end and thanks in advance for your help.
